feat: new report paradigm. Pos report timer

+++ main/u_convert.c
@@ -38,5 +38,5 @@ uint32_t mm_to_pulses_f(float mm) {
 }
 
 uint32_t pulses_to_microm(int32_t pulses) {
-    return pulses * (int32_t)settings.units.mm_rev * 1000 / (int32_t)settings.units.pulses_rev;
+    return pulses * settings.units.mm_rev * 1000 / settings.units.pulses_rev;
 }
